NP-44789 Remove redundant padding
